+++ include/qwqdsp/pitch/yin.hpp
@@ -15,19 +15,19 @@ class Yin {
 public:
     void Init(float fs, int size) {
         fs_ = fs;
-        delta_corr_.resize(size / 2);
-        powers_.resize(size / 2);
-        temp_.resize(size);
+        delta_corr_.resize(static_cast<size_t>(size) / 2);
+        powers_.resize(static_cast<size_t>(size) / 2);
+        temp_.resize(static_cast<size_t>(size));
         SetMinPitch(min_pitch_);
         SetMaxPitch(max_pitch_);
-        fft_.Init(size);
+        fft_.Init(static_cast<size_t>(size));
         fft1_.resize(fft_.NumBins());
         fft2_.resize(fft_.NumBins());
     }
 
     void Process(std::span<const float> block) noexcept {
-        int num_samples = block.size();
-        int max_tal = num_samples / 2;
+        size_t num_samples = block.size();
+        size_t max_tal = num_samples / 2;
 
         // step1 delta auto correlation(x[0:N/2] and x[0:N])
         // using x = block, len(x) = N
@@ -37,21 +37,21 @@ public:
         {
             std::fill(powers_.begin(), powers_.end(), 0.0f);
             // powers[0] = sum x[i]^2, i from 0 to N/2
-            for (int i = 0; i < max_tal; ++i) {
+            for (size_t i = 0; i < max_tal; ++i) {
                 powers_[0] += block[i] * block[i];
             }
 
             // powers[tau] = sum x[i+tau]^2 = sum x[j]^2, j from tau to N/2+tau
-            for (int tau = 1; tau < max_tal; ++tau) {
+            for (size_t tau = 1; tau < max_tal; ++tau) {
                 powers_[tau] =
                     powers_[tau - 1] - block[tau - 1] * block[tau - 1] + block[tau + max_tal] * block[tau + max_tal];
             }
 
             // fft circular correlation of x[0:N/2] and x[tau:N/2+tau]
             fft_.FFT(block, fft1_);
-            for (int i = 0; i < max_tal; ++i) {
+            for (size_t i = 0; i < max_tal; ++i) {
                 temp_[i] = block[max_tal - 1 - i]; // why reverse x[0:N/2]?
-                temp_[i + max_tal] = 0;
+                temp_[i + max_tal] = 0.0f;
             }
             fft_.FFT(temp_, fft2_);
 
@@ -63,7 +63,7 @@ public:
             }
             fft_.IFFT(temp_, fft1_);
 
-            for (int i = 0; i < max_tal; ++i) {
+            for (size_t i = 0; i < max_tal; ++i) {
                 delta_corr_[i] = powers_[0] + powers_[i] - 2 * temp_[max_tal - 1 + i]; // why reverse correlation
                 delta_corr_[i] = std::max(delta_corr_[i], 0.0f);
             }
@@ -72,11 +72,11 @@ public:
         // step2 CMNDF
         {
             float sum = 0.0f;
-            delta_corr_[0] = 1;
-            for (int tal = 1; tal < max_tal; ++tal) {
+            delta_corr_[0] = 1.0f;
+            for (size_t tal = 1; tal < max_tal; ++tal) {
                 sum += delta_corr_[tal];
                 if (sum != 0.0f) {
-                    delta_corr_[tal] *= tal / sum;
+                    delta_corr_[tal] *= static_cast<float>(tal) / sum;
                 }
                 else {
                     delta_corr_[tal] = 1.0f;
@@ -85,10 +85,11 @@ public:
         }
 
         // step3 find tau
-        int max_ifbin = std::min(max_bin_, max_tal - 1);
+        int max_ifbin = std::min(max_bin_, static_cast<int>(max_tal - 1));
         int where = -1;
         for (int i = min_bin_; i < max_ifbin; ++i) {
-            if (delta_corr_[i] < threshold_ && delta_corr_[i] < delta_corr_[i + 1]) {
+            if (delta_corr_[static_cast<size_t>(i)] < threshold_
+                && delta_corr_[static_cast<size_t>(i)] < delta_corr_[static_cast<size_t>(i) + 1]) {
                 where = i;
                 break;
             }
@@ -96,24 +97,24 @@ public:
         if (where == -1) {
             float min = delta_corr_.front();
             for (int i = min_bin_; i < max_ifbin; ++i) {
-                if (delta_corr_[i] < min) {
-                    min = delta_corr_[i];
+                if (delta_corr_[static_cast<size_t>(i)] < min) {
+                    min = delta_corr_[static_cast<size_t>(i)];
                     where = i;
                 }
             }
         }
 
         // step4 parabola interpolation
-        float preiod = where;
-        if (where > 0 && where < max_tal - 1) {
-            float s0 = delta_corr_[where - 1];
-            float s1 = delta_corr_[where];
-            float s2 = delta_corr_[where + 1];
+        float preiod = static_cast<float>(where);
+        if (where > 0 && where < static_cast<int>(max_tal) - 1) {
+            float s0 = delta_corr_[static_cast<size_t>(where - 1)];
+            float s1 = delta_corr_[static_cast<size_t>(where)];
+            float s2 = delta_corr_[static_cast<size_t>(where + 1)];
             if (s1 < s0 && s1 < s2) {
                 float frac = 0.5f * (s2 - s0) / (2.0f * s1 - s2 - s0 + 1e-18f);
-                preiod = where + frac;
+                preiod = static_cast<float>(where) + frac;
                 pitch_.pitch_hz = fs_ / preiod;
-                pitch_.non_period_ratio = delta_corr_[where];
+                pitch_.non_period_ratio = delta_corr_[static_cast<size_t>(where)];
             }
             else {
                 // 无峰值，大概是噪声或者在外面吧
@@ -132,12 +133,12 @@ public:
 
     void SetMinPitch(float min_val) noexcept {
         min_pitch_ = min_val;
-        max_bin_ = std::round(fs_ / min_val);
+        max_bin_ = static_cast<int>(std::round(fs_ / min_val));
     }
 
     void SetMaxPitch(float max_val) noexcept {
         max_pitch_ = max_val;
-        min_bin_ = std::round(fs_ / max_val);
+        min_bin_ = static_cast<int>(std::round(fs_ / max_val));
     }
 
     void SetThreshold(float thredshold) noexcept {
