+++ include/qwqdsp/window/hamming.hpp
@@ -16,13 +16,13 @@ struct Hamming {
         const size_t N = x.size();
         if (for_analyze_not_fir) {
             for (size_t n = 0; n < N; ++n) {
-                const float t = n / static_cast<float>(N) - 0.5f;
+                const float t = static_cast<float>(n) / static_cast<float>(N) - 0.5f;
                 x[n] = 0.53836f + 0.46164f * std::cos(std::numbers::pi_v<float> * 2 * t);
             }
         }
         else {
             for (size_t n = 0; n < N; ++n) {
-                const float t = n / (N - 1.0f) - 0.5f;
+                const float t = static_cast<float>(n) / static_cast<float>(N - 1) - 0.5f;
                 x[n] = 0.53836f + 0.46164f * std::cos(std::numbers::pi_v<float> * 2 * t);
             }
         }
@@ -32,13 +32,13 @@ struct Hamming {
         const size_t N = x.size();
         if (for_analyze_not_fir) {
             for (size_t n = 0; n < N; ++n) {
-                const float t = n / static_cast<float>(N) - 0.5f;
+                const float t = static_cast<float>(n) / static_cast<float>(N) - 0.5f;
                 x[n] *= 0.53836f + 0.46164f * std::cos(std::numbers::pi_v<float> * 2 * t);
             }
         }
         else {
             for (size_t n = 0; n < N; ++n) {
-                const float t = n / (N - 1.0f) - 0.5f;
+                const float t = static_cast<float>(n) / static_cast<float>(N - 1) - 0.5f;
                 x[n] *= 0.53836f + 0.46164f * std::cos(std::numbers::pi_v<float> * 2 * t);
             }
         }
@@ -47,7 +47,7 @@ struct Hamming {
     static void DWindow(std::span<float> x) noexcept {
         const size_t N = x.size();
         for (size_t n = 0; n < N; ++n) {
-            const float t = n / static_cast<float>(N) - 0.5f;
+            const float t = static_cast<float>(n) / static_cast<float>(N) - 0.5f;
             x[n] = -0.46164f * std::numbers::pi_v<float> * 2 * std::sin(std::numbers::pi_v<float> * 2 * t);
         }
     }
