+++ include/qwqdsp/window/hann.hpp
@@ -16,14 +16,14 @@ struct Hann {
         const size_t N = x.size();
         if (for_analyze_not_fir) {
             for (size_t n = 0; n < N; ++n) {
-                const float t = n / static_cast<float>(N);
-                x[n] = 0.5 * (1.0 - cos(2.0 * std::numbers::pi_v<float> * t));
+                const float t = static_cast<float>(n) / static_cast<float>(N);
+                x[n] = 0.5f * (1.0f - std::cos(2.0f * std::numbers::pi_v<float> * t));
             }
         }
         else {
             for (size_t n = 0; n < N; ++n) {
-                const float t = n / (N - 1.0f);
-                x[n] = 0.5 * (1.0 - cos(2.0 * std::numbers::pi_v<float> * t));
+                const float t = static_cast<float>(n) / static_cast<float>(N - 1);
+                x[n] = 0.5f * (1.0f - std::cos(2.0f * std::numbers::pi_v<float> * t));
             }
         }
     }
@@ -32,14 +32,14 @@ struct Hann {
         const size_t N = x.size();
         if (for_analyze_not_fir) {
             for (size_t n = 0; n < N; ++n) {
-                const float t = n / static_cast<float>(N);
-                x[n] *= 0.5 * (1.0 - cos(2.0 * std::numbers::pi_v<float> * t));
+                const float t = static_cast<float>(n) / static_cast<float>(N);
+                x[n] *= 0.5f * (1.0f - std::cos(2.0f * std::numbers::pi_v<float> * t));
             }
         }
         else {
             for (size_t n = 0; n < N; ++n) {
-                const float t = n / (N - 1.0f);
-                x[n] *= 0.5 * (1.0 - cos(2.0 * std::numbers::pi_v<float> * t));
+                const float t = static_cast<float>(n) / static_cast<float>(N - 1);
+                x[n] *= 0.5f * (1.0f - std::cos(2.0f * std::numbers::pi_v<float> * t));
             }
         }
     }
@@ -47,7 +47,7 @@ struct Hann {
     static void DWindow(std::span<float> x) noexcept {
         const size_t N = x.size();
         for (size_t n = 0; n < N; ++n) {
-            const float t = n / static_cast<float>(N);
+            const float t = static_cast<float>(n) / static_cast<float>(N);
             x[n] = std::numbers::pi_v<float> * std::sin(std::numbers::pi_v<float> * 2 * t);
         }
     }
